% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{panel.hanoi}
\alias{panel.hanoi}
\title{Custom Panel Function for Histograms in Lattice Plots}
\usage{
panel.hanoi(x, y, horizontal, breaks = "Sturges", ...)
}
\description{
Utility functions used in lattice plots to show distribution overlays in bwplot comparisons.
}
\keyword{internal}
