% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_stage_crm_logistic.R
\name{two_stage_crm_logistic}
\alias{two_stage_crm_logistic}
\title{C-optimal based two-stage Continual Reassessment Method (CRM) Simulation for logistic model with 2 parameters}
\usage{
two_stage_crm_logistic(
  p0 = 0.4,
  theta = c(-3, 2),
  theta_0 = c(-3.1, 1.8),
  N = 24,
  n_initial = 3,
  q_0 = 0.05,
  q_2 = 0.4,
  q_1 = 0.9,
  lim_sup_prob = 0.7,
  show_plot = FALSE,
  seed = 1234
)
}
\arguments{
\item{p0}{Target toxicity probability. Default is 0.4.}

\item{theta}{True value for the vector of parameters for the dose-toxicity curve. Default is c(-3,2).}

\item{theta_0}{Nominal value of the vector of parameters. Default is c(-3.1, 1.8).}

\item{N}{Total number of patients (including both stages). Default is 24.}

\item{n_initial}{Number of patients per dose in stage 1. Default is 3.}

\item{q_0}{Toxicity probability at first dose. Default is 0.02.}

\item{q_2}{Fraction of patients in stage 1. Default is 0.5.}

\item{q_1}{Target probability of observing at least one toxicity during stage 1. Default is 0.9.}

\item{show_plot}{Logical. If TRUE, plots the dose levels for patients in the trial. Default is FALSE.}

\item{seed}{Random seed for reproducibility. Default is 1234.}
}
\value{
A list with:
\describe{
  \item{n_toxicities}{Total number of toxicities observed.}
  \item{mtd_estimated}{Estimated Maximum Tolerated Dose (MTD). If no toxicity is observed in the first stage, the MTD is set to the largest dose level used.}
  \item{mle_theta}{Vector of estimated parameters \eqn{\theta = (\alpha, \beta)} for the logistic dose-toxicity model. Set to \code{NA} if no toxicity is detected in stage 1.}
  \item{x}{Vector of dose levels administered.}
  \item{y}{Vector of toxicity outcomes (1 = toxic, 0 = non-toxic).}
}
}
\description{
Performs one simulation run of a two-stage CRM design for estimating the Maximum Tolerated Dose (MTD) in phase I trials.
}
\note{
If no toxicity is observed during the first stage of the trial (i.e., \code{sum(y) == 0}), the simulation is terminated.
A warning is issued, and the MTD is conservatively estimated as the highest dose level reached.
The value of \code{mle_theta} is set to \code{NA} in this case.
}
\examples{
result <- two_stage_crm_logistic(show_plot = TRUE)
print(result$mtd_estimated)

}
